\name{as.SO3}
\alias{as.SO3}
\title{Add SO3 class to object}
\usage{
  as.SO3(x)
}
\arguments{
  \item{x}{an R object}
}
\value{
  x with class "SO3"
}
\description{
  Add SO3 class to object
}
\seealso{
  Other SO3: \code{\link{id.SO3}}, \code{\link{is.SO3}},
  \code{\link{SO3}}, \code{\link{SO3-class}},
  \code{\link{SO3.default}}, \code{\link{SO3.Q4}}
}

