\docType{data}
\name{id.Q4}
\alias{id.Q4}
\title{Identity Quaternion}
\format{Q4 [1, 1:4] 1 0 0 0}
\usage{
  id.Q4
}
\description{
  This is a Q4 object that represents the identity rotation
  for quaternions. More specifically
  \deqn{q=[1,0,0,0]^\top}{q=[1,0,0,0]'} represnents the
  roation of some coordinate axis about any axis through
  the angle 0.
}
\seealso{
  Other Q4: \code{\link{as.Q4}}, \code{\link{is.Q4}},
  \code{\link{Q4}}, \code{\link{Q4-class}},
  \code{\link{Q4.default}}, \code{\link{Q4.SO3}}
}
\keyword{datasets}

