\name{is.Q4}
\alias{is.Q4}
\title{A function to determine if a given object is in unit quaternion or not.}
\usage{
  is.Q4(x)
}
\arguments{
  \item{x}{numeric 1-by-4 vector of length 4}
}
\value{
  logical T if the vector is a unit quaternion and false
  otherwise
}
\description{
  A function to determine if a given object is in unit
  quaternion or not.
}
\seealso{
  Other Q4: \code{\link{as.Q4}}, \code{\link{id.Q4}},
  \code{\link{Q4}}, \code{\link{Q4-class}},
  \code{\link{Q4.default}}, \code{\link{Q4.SO3}}
}

